CharTuple

+++ src/OrleanSpaces/Tuples/Typed/CharTuple.cs
@@ -26,7 +26,17 @@ public readonly struct CharTuple : IValueTuple<char, CharTuple>, ISpanFormattabl
 
     public bool Equals(CharTuple other)
     {
-        NumericMarshaller<char, byte> marshaller = new(fields.AsSpan(), other.fields.AsSpan());
+        // Since 'char' is not a number (doesn't implement INumber<T>), we are transforming it into a type which does implement INumber<T>.
+        // The sizeof(char) = 2 Bytes, therefor it can be represented by many number types, but the lowest possible (the one that provides the best parallelization)
+        // number type that can fully represent any type of 'char', is 'short/ushort'.
+
+        // In systems where 128-bit vector operations are subject to hardware acceleration, a total of 8 operations can be performed on 'ushort's
+        //      128 bits / 2 Bytes = 128 bits / 16 bits =  8 bits --> means: we can compare 8 chars at the same time!
+
+        // In systems where 256-bit vector operations are subject to hardware acceleration, a total of 16 operations can be performed on 'ushort's
+        //      256 bits / 2 Bytes = 256 bits / 16 bits =  16 bits --> means: we can compare 16 chars at the same time!
+
+        NumericMarshaller<char, ushort> marshaller = new(fields.AsSpan(), other.fields.AsSpan());
         return marshaller.TryParallelEquals(out bool result) ? result : this.SequentialEquals(other);
     }
